split backend_gktagg.py in to two parts
_backend_gtkagg.py has no dependence on `Gcf`, backend_gtkagg.py
depends on `Gcf` and maintains reverse compatibility.
